Add files via upload
🔍 Exploring Historical COVID-19 Data: A Tableau Project 📈

As we navigate the complexities of the ongoing pandemic in 2024, I’m excited to share a comprehensive analysis of COVID-19 data from 2019, visualized using Tableau. This project provides valuable insights into the early stages of the pandemic, including:

🔹 Age Group Distribution
🔹 State-Wise Vaccine Coverage
🔹 Hospital Beds & Medical Facilities
🔹 Death & Recovery Statistics
🔹 ICMR Testing Labs & Statewise Testing Data
🔹 Individual and Population Demographics

While this data reflects a past moment, it offers critical context for understanding the pandemic’s progression and the evolution of public health strategies. Explore the interactive dashboard to see how the early data compares to the current situation.